% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{dendrogram.fakeCommunities}
\alias{dendrogram.fakeCommunities}
\title{Returns pre-calculated dendrogram}
\usage{
dendrogram.fakeCommunities(obj, ...)
}
\arguments{
\item{obj}{fakeCommunities object}

\item{...}{dropped}
}
\value{
dendrogram
}
\description{
Returns pre-calculated dendrogram
}
\examples{
rLeidenComm = suppressWarnings(rleiden.community(conosGraph, n.cores=1))
dendrogram.fakeCommunities(rLeidenComm)
}
