% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R, R/communities.R
\name{find_partition_rcpp}
\alias{find_partition_rcpp}
\alias{find_partition}
\title{Finds the optimal partition using the Leiden algorithm}
\usage{
find_partition_rcpp(
  edgelist,
  edgelist_length,
  num_vertices,
  direction,
  edge_weights,
  resolution = 1,
  niter = 2L,
  nrep = 1L,
  seed = 1L
)

find_partition(
  graph,
  edge_weights,
  resolution = 1,
  niter = 2,
  nrep = 1,
  seed = 1
)
}
\arguments{
\item{edgelist}{The graph edge list}

\item{edgelist_length}{integer The length of the graph edge list}

\item{num_vertices}{integer The number of vertices in the graph}

\item{direction}{boolean Whether the graph is directed or undirected}

\item{edge_weights}{Vector of edge weights. In weighted graphs, a real number is assigned to each (directed or undirected) edge. For an unweighted graph, this is set to 1. Refer to igraph, weighted graphs.}

\item{resolution}{Integer resoluiton parameter controlling communities detected (default=1.0) Higher resolutions lead to more communities, while lower resolutions lead to fewer communities.}

\item{niter}{Number of iterations that the algorithm should be run for (default=2)}

\item{nrep}{Number of random replicate starts with random number being updated. (default=10) The result with the best modularity will be returned.}

\item{seed}{Initial random seed for reproducible result. (default=1) When running multiple replicates, the seed increments by 1 every new start.}

\item{graph}{The igraph graph to define the partition on}
}
\value{
A vector of membership values
}
\description{
Finds the optimal partition using the Leiden algorithm
}
\details{
For notes of the graph object, refer to https://igraph.org/c/doc/igraph-Basic.html
}
\examples{
library(igraph)
library(leidenAlg)

g <- make_star(10)
E(g)$weight <- seq(ecount(g))
find_partition(g, E(g)$weight)

}
