% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{getClusterGraph}
\alias{getClusterGraph}
\title{Collapse vertices belonging to each cluster in a graph}
\usage{
getClusterGraph(
  graph,
  groups,
  method = "sum",
  plot = FALSE,
  seed = 1,
  node.scale = 50,
  edge.scale = 50,
  edge.alpha = 0.3,
  ...
)
}
\arguments{
\item{graph}{igraph object Graph to be clustered}

\item{groups}{Factor on vertices describing cluster assignment (can specify integer vertex ids, or character vertex names which will be matched)}

\item{method}{string Method used, either "sum" or "paga" (default="sum"). "paga" refers to PAGA, <https://github.com/theislab/paga>}

\item{plot}{boolean Whether to show collapsed graph plot (default=FALSE)}

\item{seed}{numeric Set seed via set.seed() for plotting (default=1)}

\item{node.scale}{numeric Value to scale the size of the nodes, via 'vertex.size' in igraph::plot() (default=50)}

\item{edge.scale}{numeric Value to scale the edge thickness of the noes, via 'edge.width' in graph::plot()  (default=50)}

\item{edge.alpha}{numeric Value to scale the opacity of 'alpha.f' in adjustcolor(); typically in [0,1] (default=0.3)}

\item{...}{arguments passed to collapseGraphSum()}
}
\value{
collapsed graph
}
\description{
Collapse vertices belonging to each cluster in a graph
}
