% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{leiden_community}
\alias{leiden_community}
\title{find the optimal graph partition using the Leiden algorithm, Traag et al 2019 https://doi.org/10.1038/s41598-019-41695-z. Refer to find_partition() in the python interface, https://github.com/vtraag/leidenalg}
\usage{
leiden_community(graph, edge_weights, resolution = 1, niter = 2L)
}
\arguments{
\item{graph}{The input graph for which to detect communities.}

\item{edge_weights}{Weights of edges. Can be either an iterable or an edge attribute.}

\item{resolution}{numeric}

\item{niter}{integer Number of iterations to run the Leiden algorithm. By default, 2 iterations are run. If the number of iterations is negative, the Leiden algorithm is run until an iteration in which there was no improvement.}
}
\description{
find the optimal graph partition using the Leiden algorithm, Traag et al 2019 https://doi.org/10.1038/s41598-019-41695-z. Refer to find_partition() in the python interface, https://github.com/vtraag/leidenalg
}
