% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/communities.R
\name{set.names}
\alias{set.names}
\title{setNames wrapper function, also called 'sn' elsewhere}
\usage{
set.names(x)
}
\arguments{
\item{x}{an object for which names attribute will be meaningful}
}
\value{
An object with names assigned equal to values
}
\description{
Set names equal to values
}
\examples{
vec = c(1, 2, 3, 4)
set.names(vec)
}
\keyword{internal}
